refactor(ui-library): cos text area
